[Serve] Move kill-heavy tests to isolated file with function-scoped fixtures
Move 5 tests that kill deployment actors or the controller into
test_deployment_actors_recovery.py with a function-scoped fixture
(fresh Ray cluster per test). This prevents accumulated controller
state from causing GCS degradation and cascading failures.

Tests moved:
- test_deployment_actor_restarts_on_crash
- test_deployment_actor_health_check_failure_then_recovery_to_healthy
- test_cached_get_deployment_actor_handle_stale_after_recreation
- test_deployment_actor_survives_controller_restart
- test_controller_restart_preserves_mutated_actor_state

Reverts test_gcs_failure.py to master (controller-restart tests now
in the new file instead).
