+++ python/ray/serve/tests/test_deployment_actors.py
@@ -10,7 +10,6 @@ import yaml
 import ray
 from ray import serve
 from ray._common.test_utils import SignalActor, wait_for_condition
-from ray.exceptions import RayActorError
 from ray.serve._private.common import DeploymentID, DeploymentStatus, ReplicaState
 from ray.serve._private.constants import (
     REPLICA_HEALTH_CHECK_UNHEALTHY_THRESHOLD,
@@ -71,7 +70,6 @@ class ConstructorGate:
         self._open = True
 
     def wait_until_open(self):
-        import time
 
         while not self._open:
             time.sleep(0.01)
@@ -796,271 +794,6 @@ def test_redeployment_with_no_actors_cleans_up_old(serve_instance):
     wait_for_condition(lambda: httpx.get(url).text == "v2")
 
     wait_for_condition(_check_no_deployment_actors, timeout=15)
-
-
-def test_deployment_actor_restarts_on_crash(serve_instance):
-    """When a deployment actor dies, the Serve controller recreates it (no Ray auto-restart).
-
-    Replicas call ``serve.get_deployment_actor()`` per request here, so they resolve a
-    current handle after recreation. HTTP bodies include ``ray_actor_id`` so we assert a
-    new id after Serve recreates the deployment actor.
-    """
-
-    @serve.deployment(
-        name="CrashTestDeployment",
-        deployment_actors=[
-            DeploymentActorConfig(
-                name="counter",
-                actor_class=SharedCounter,
-                init_kwargs={"start": 100},
-            ),
-        ],
-    )
-    class CrashTestDeployment:
-        def __call__(self):
-            counter = serve.get_deployment_actor("counter")
-            val = ray.get(counter.get.remote())
-            aid = ray.get(counter.ray_actor_id.remote())
-            return f"{val},{aid}"
-
-    serve.run(CrashTestDeployment.bind())
-    url = f"{get_application_url()}/"
-
-    def _parse_val_actor_id_crash(text: str) -> tuple[str, str]:
-        val, aid = text.split(",", 1)
-        return val, aid
-
-    wait_for_condition(
-        lambda: _parse_val_actor_id_crash(httpx.get(url).text)[0] == "100"
-    )
-    _, old_actor_id = _parse_val_actor_id_crash(httpx.get(url).text)
-
-    actor_names = [
-        n
-        for n in _get_deployment_actor_names()
-        if "CrashTestDeployment" in n and "counter" in n
-    ]
-    assert len(actor_names) == 1
-    handle = ray.get_actor(actor_names[0], namespace=SERVE_NAMESPACE)
-
-    ray.kill(handle, no_restart=True)
-
-    def recovered_new_deployment_actor():
-        text = httpx.get(url, timeout=5).text
-        val, aid = _parse_val_actor_id_crash(text)
-        return val == "100" and aid != old_actor_id
-
-    wait_for_condition(recovered_new_deployment_actor, timeout=120)
-
-
-def test_deployment_actor_health_check_failure_then_recovery_to_healthy(serve_instance):
-    """E2E: deployment actor fails controller health poll → UNHEALTHY → recreate → HEALTHY.
-
-    The controller health-checks deployment-scoped actors with ``__ray_ready__``. If the
-    worker is gone, ``ray.get`` raises ``RayActorError`` and the deployment is marked
-    UNHEALTHY immediately; the controller then creates a new actor (same as the
-    repeated-timeout/APP_FAILURE path, but without waiting for the unhealthy threshold).
-
-    After killing the deployment actor, we ``close()`` a :class:`ConstructorGate` so the
-    replacement actor blocks in ``__init__``. While it is stuck starting, Serve cannot
-    return to HEALTHY (``_deployment_actors_satisfied_for_target`` requires RUNNING), so
-    ``DeploymentStatus.UNHEALTHY`` is stable and easy to assert without racing the status
-    API. Opening the gate lets construction finish and the deployment recover.
-    """
-
-    dep_name = "DAHealthRecoveryDeployment"
-    logical_actor = "counter"
-    gate = ConstructorGate.remote()
-
-    @serve.deployment(
-        name=dep_name,
-        deployment_actors=[
-            DeploymentActorConfig(
-                name=logical_actor,
-                actor_class=GatedSharedCounter,
-                init_kwargs={"gate": gate, "start": 42},
-            ),
-        ],
-    )
-    class DAHealthRecoveryDeployment:
-        def __call__(self):
-            counter = serve.get_deployment_actor(logical_actor)
-            val = ray.get(counter.get.remote())
-            aid = ray.get(counter.ray_actor_id.remote())
-            return f"{val},{aid}"
-
-    serve.run(DAHealthRecoveryDeployment.bind(), name=SERVE_DEFAULT_APP_NAME)
-    url = f"{get_application_url()}/"
-
-    def _parse_val_actor_id_health(text: str) -> tuple[str, str]:
-        val, aid = text.split(",", 1)
-        return val, aid
-
-    def deployment_status():
-        return (
-            serve.status()
-            .applications[SERVE_DEFAULT_APP_NAME]
-            .deployments[dep_name]
-            .status
-        )
-
-    def counter_actor_name() -> str:
-        names = [
-            n
-            for n in _get_deployment_actor_names()
-            if dep_name in n and logical_actor in n
-        ]
-        assert len(names) == 1
-        return names[0]
-
-    def counter_actor_id_for_name(name: str) -> str:
-        for a in list_actors(filters=[("state", "=", "ALIVE")]):
-            if a.get("name") == name:
-                return a.actor_id
-        raise AssertionError(f"No ALIVE actor record for name={name!r}")
-
-    wait_for_condition(
-        lambda: _parse_val_actor_id_health(httpx.get(url).text)[0] == "42"
-    )
-    assert deployment_status() == DeploymentStatus.HEALTHY
-
-    da_name = counter_actor_name()
-    old_actor_id = counter_actor_id_for_name(da_name)
-    _, http_actor_id_before = _parse_val_actor_id_health(httpx.get(url).text)
-    assert http_actor_id_before == old_actor_id
-
-    ray.get(gate.close.remote())
-    ray.kill(
-        ray.get_actor(da_name, namespace=SERVE_NAMESPACE),
-        no_restart=True,
-    )
-
-    wait_for_condition(
-        lambda: deployment_status() == DeploymentStatus.UNHEALTHY,
-        timeout=120,
-    )
-    assert deployment_status() == DeploymentStatus.UNHEALTHY
-
-    ray.get(gate.open.remote())
-
-    wait_for_condition(
-        lambda: deployment_status() == DeploymentStatus.HEALTHY,
-        timeout=120,
-    )
-
-    def http_shows_new_deployment_actor():
-        val, aid = _parse_val_actor_id_health(httpx.get(url, timeout=5).text)
-        return val == "42" and aid != old_actor_id
-
-    wait_for_condition(http_shows_new_deployment_actor, timeout=120)
-
-    new_name = counter_actor_name()
-    assert new_name == da_name
-    new_actor_id = counter_actor_id_for_name(new_name)
-    assert new_actor_id != old_actor_id
-    _, http_actor_id_after = _parse_val_actor_id_health(httpx.get(url).text)
-    assert http_actor_id_after == new_actor_id
-
-
-def test_cached_get_deployment_actor_handle_stale_after_recreation(serve_instance):
-    """Stale handle in ``__init__`` vs cache + refresh after the deployment actor dies.
-
-    (1) A handle from ``get_deployment_actor`` stored only in ``__init__`` still
-    points at the dead actor after Serve recreates the named actor; HTTP stays 5xx.
-    Logs typically show ``ActorDiedError`` / ``RayActorError`` from ``ray.get`` on
-    that stale handle.
-
-    (2) **Recommended** — on ``RayActorError`` from ``ray.get``, call
-    ``serve.get_deployment_actor`` again. That lookup can raise ``ValueError`` with a
-    message that lists three causes (including namespace); after a kill the usual
-    case is that the name is not registered yet—retry briefly until the controller
-    finishes recreating the actor. Alternatively resolve the actor on every request;
-    see ``test_deployment_actor_restarts_on_crash``.
-
-    Expected log noise: replica ``ERROR``/``500`` lines while the actor is missing or
-    stale are normal for phase (1); phase (2) should stay quiet once refresh retries
-    ``get_deployment_actor``.
-    """
-
-    counter_cfg = DeploymentActorConfig(
-        name="counter",
-        actor_class=SharedCounter,
-        init_kwargs={"start": 7},
-    )
-
-    def kill_stale_test_counter():
-        names = [
-            n
-            for n in _get_deployment_actor_names()
-            if "StaleHandleDeployment" in n and "counter" in n
-        ]
-        assert len(names) == 1
-        ray.kill(
-            ray.get_actor(names[0], namespace=SERVE_NAMESPACE),
-            no_restart=True,
-        )
-
-    @serve.deployment(
-        name="StaleHandleDeployment",
-        deployment_actors=[counter_cfg],
-    )
-    class StaleHandleDeployment:
-        def __init__(self):
-            self._counter = serve.get_deployment_actor("counter")
-
-        def __call__(self):
-            return str(ray.get(self._counter.get.remote()))
-
-    serve.run(StaleHandleDeployment.bind())
-    url = f"{get_application_url()}/"
-    wait_for_condition(lambda: httpx.get(url).text == "7")
-
-    kill_stale_test_counter()
-
-    def replica_still_fails_after_recreation():
-        r = httpx.get(url, timeout=10)
-        return r.status_code >= 500
-
-    wait_for_condition(replica_still_fails_after_recreation, timeout=120)
-
-    @serve.deployment(
-        name="StaleHandleDeployment",
-        deployment_actors=[counter_cfg],
-    )
-    class CacheAndRefreshOnRayActorError:
-        def __init__(self):
-            self._counter = serve.get_deployment_actor("counter")
-
-        def _resolve_counter_after_actor_died(self):
-            # After Serve kills the deployment actor, the same name may not exist in
-            # GCS until recreation completes; ray.get_actor raises ValueError (message
-            # also mentions namespace as a generic possibility).
-            deadline = time.monotonic() + 30.0
-            last_exc = None
-            while time.monotonic() < deadline:
-                try:
-                    self._counter = serve.get_deployment_actor("counter")
-                    return
-                except ValueError as e:
-                    last_exc = e
-                    time.sleep(0.05)
-            if last_exc is not None:
-                raise last_exc
-            raise TimeoutError(
-                "Timed out waiting for deployment actor name after recreation."
-            )
-
-        def __call__(self):
-            try:
-                return str(ray.get(self._counter.get.remote()))
-            except RayActorError:
-                self._resolve_counter_after_actor_died()
-                return str(ray.get(self._counter.get.remote()))
-
-    serve.run(CacheAndRefreshOnRayActorError.bind())
-    wait_for_condition(lambda: httpx.get(url).text == "7")
-    kill_stale_test_counter()
-    wait_for_condition(lambda: httpx.get(url).text == "7", timeout=120)
 
 
 def test_deployment_actor_constructor_failure_app_status(serve_instance):
